# docs/swagger.yaml
definitions:
  db.AllDraft:
    properties:
      author_name:
        type: string
      created_at:
        type: integer
      draft_id:
        type: string
      interest:
        items:
          type: string
        type: array
      tagline:
        type: string
      title_data:
        type: string
    type: object
  db.DraftDB:
    properties:
      created_at:
        type: integer
      draft_id:
        type: string
      interest:
        items:
          type: string
        type: array
      is_published:
        type: boolean
      post_data:
        $ref: '#/definitions/models.JSONString'
      preview_image:
        type: string
      tagline:
        type: string
      user_id:
        type: string
    type: object
  db.Interest:
    properties:
      name:
        type: string
    type: object
  golaerror.Error:
    properties:
      additionalData:
        $ref: '#/definitions/golaerror.additionalData'
      errorCode:
        type: string
      errorMessage:
        type: string
    type: object
  golaerror.additionalData:
    type: object
  models.GetAllDraftRequest:
    properties:
      limit:
        type: integer
      start_value:
        type: integer
      user_id:
        type: string
    required:
    - limit
    - start_value
    - user_id
    type: object
  models.JSONString:
    properties:
      types.JSONText:
        items:
          type: integer
        type: array
    type: object
  models.UpsertDraft:
    properties:
      draft_id:
        type: string
      post_data:
        $ref: '#/definitions/models.JSONString'
      user_id:
        type: string
    required:
    - draft_id
    - user_id
    type: object
...
  request.DraftURIRequest:
    properties:
      draftID:
        type: string
    required:
    - draftID
    type: object
  request.InterestsSaveRequest:
    properties:
      draft_id:
        type: string
      interest:
        type: string
      user_id:
        type: string
    required:
    - draft_id
    - interest
    - user_id
    type: object
...
  request.PreviewImageSaveRequest:
    properties:
      draft_id:
        type: string
      preview_image:
        type: string
      user_id:
        type: string
    required:
    - draft_id
    - preview_image
    - user_id
    type: object
...
  request.SearchInterests:
    properties:
      searchKeyword:
        type: string
      selectedTags:
        items:
          type: string
        type: array
    required:
    - selectedTags
    type: object
  request.TaglineSaveRequest:
    properties:
      draft_id:
        type: string
      tagline:
        type: string
      user_id:
        type: string
    required:
    - draft_id
    - tagline
    - user_id
    type: object
...
info:
  contact: {}
paths:
  /api/post/v1/draft/delete-interest:
    post:
      consumes:
      - application/json
      description: delete interest for draft
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.InterestsSaveRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: DeleteInterest
      tags:
      - draft
  /api/post/v1/draft/delete/:draft_id:
    post:
      consumes:
      - application/json
      description: delete draft for a user
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DraftURIRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: DeleteDraft
      tags:
      - draft
  /api/post/v1/draft/get-all-draft:
    post:
      consumes:
      - application/json
      description: get all draft for given draft id
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GetAllDraftRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.AllDraft'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetAllDraft
      tags:
      - draft
  /api/post/v1/draft/get-draft/:draft_id:
    get:
      consumes:
      - application/json
      description: get draft for given draft id
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DraftURIRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.DraftDB'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetDraft
      tags:
      - draft
  /api/post/v1/draft/preview-draft/:draft_id:
    get:
      consumes:
      - application/json
      description: get preview draft for a given draft id
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DraftURIRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetPreviewDraft
      tags:
      - draft
  /api/post/v1/draft/tagline:
    post:
      consumes:
      - application/json
      description: save or update tagline for draft
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TaglineSaveRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: SaveTagline
      tags:
      - draft
  /api/post/v1/draft/upsert-interests:
    post:
      consumes:
      - application/json
      description: save or update interest for draft
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.InterestsSaveRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: SaveInterests
      tags:
      - draft
  /api/post/v1/draft/upsert-preview-image:
    post:
      consumes:
      - application/json
      description: saves preview image for draft
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PreviewImageSaveRequest'
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: SavePreviewImage
      tags:
      - draft
  /api/post/v1/draft/upsertDraft:
    post:
      consumes:
      - application/json
      description: Save new draft or update existing draft
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpsertDraft'
      responses:
        "200":
          description: ""
        "400":
          description: ""
        "500":
          description: ""
      summary: SaveDraft
      tags:
      - draft
...
  /api/post/v1/post/comment:
    post:
      consumes:
      - application/json
      description: get all interests or search a interest with keyword
      parameters:
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.SearchInterests'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Interest'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/golaerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/golaerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/golaerror.Error'
      summary: GetInterests
      tags:
      - interest
...
swagger: "2.0"